\name{Rfringe-internal}
\title{Rfringe internal functions}
\alias{aboutrfringe}
\alias{addtoproject}
\alias{analysisinfo}
\alias{autotrace}
\alias{autotrace.options}
\alias{batchadd}
\alias{changeint}
\alias{changeproject}
\alias{circlepars}
\alias{clearint}
\alias{clearproject}
\alias{closeRfringe}
\alias{closeRfringeandr}
\alias{editfringe}
\alias{fitzernikes}
\alias{helpbox}
\alias{helprfringe}
\alias{imageinfo}
\alias{maxorder}
\alias{newproject}
\alias{obstructpars}
\alias{openint}
\alias{opensavedint}
\alias{opensavedproject}
\alias{pdfreport}
\alias{pdfreport.project}
\alias{plotcontour}
\alias{plotcontour.project}
\alias{plotfringes}
\alias{plotresiduals}
\alias{plotspm.project}
\alias{plotstartest}
\alias{plotstartest.project}
\alias{plotwf}
\alias{plotwf.project}
\alias{printdetails}
\alias{printdetails.project}
\alias{printsummary}
\alias{printsummary.project}
\alias{printtoc.project}
\alias{prompts}
\alias{removefromproject}
\alias{saveint}
\alias{saveproject}
\alias{synthint}
\alias{targetconicinfo}
\alias{tclvar}
\alias{wf3d}
\alias{wf3d.project}
\alias{wf3d.vanilla}
\alias{wf3d.vanilla.project}
\description{
  Internal functions called by the \code{\link{Rfringe}} GUI.
}
\usage{
aboutrfringe()
addtoproject()
analysisinfo()
autotrace()
autotrace.options()
batchadd()
changeint()
changeproject()
circlepars()
clearint()
clearproject()
closeRfringe()
closeRfringeandr()
editfringe()
fitzernikes()
helpbox(string)
helprfringe()
imageinfo()
maxorder()
newproject()
obstructpars()
openint()
opensavedint()
opensavedproject()
pdfreport()
pdfreport.project()
plotcontour()
plotcontour.project()
plotfringes()
plotresiduals()
plotspm.project()
plotstartest()
plotstartest.project()
plotwf()
plotwf.project()
printdetails()
printdetails.project()
printsummary()
printsummary.project()
printtoc.project()
prompts(string)
removefromproject()
saveint()
saveproject()
synthint()
targetconicinfo()
tclvar()
wf3d()
wf3d.project()
wf3d.vanilla()
wf3d.vanilla.project()
}
\details{
  These are called by \code{Rfringe()} and are not to be called directly by the user.
  From the command line there are always functions stored within \code{\link{interferogram}},
  \code{\link{project}}, or \R to perform the actions carried out by these functions.
}
\author{M.L. Peck \email{mpeck1@ix.netcom.com}}
\seealso{
  \code{\link{Rfringe}},
  \code{\link{interferogram}},
  \code{\link{project}}}.
}
\keyword{internal}
